+++ analysis/filter/EdgeFilter/src/test/kotlin/de/maibornwolff/codecharta/filter/edgefilter/ParserDialogTest.kt
@@ -5,13 +5,11 @@ import com.github.kinquirer.components.promptInput
 import io.mockk.every
 import io.mockk.mockkStatic
 import io.mockk.unmockkAll
-import org.assertj.core.api.Assertions
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.AfterAll
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.TestInstance
 import picocli.CommandLine
-import java.io.File
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 class ParserDialogTest {
@@ -30,7 +28,8 @@ class ParserDialogTest {
 
         val parserArguments = ParserDialog.collectParserArgs()
 
-        Assertions.assertThat(parserArguments).isEqualTo(listOf("sampleFile.cc.json", "--output-file=sampleOutputFile", "--path-separator=/"))
+        assertThat(parserArguments)
+            .isEqualTo(listOf("sampleFile.cc.json", "--output-file=sampleOutputFile", "--path-separator=/"))
     }
 
     @Test
@@ -43,7 +42,7 @@ class ParserDialogTest {
         val parserArguments = ParserDialog.collectParserArgs()
         val cmdLine = CommandLine(EdgeFilter())
         val parseResult = cmdLine.parseArgs(*parserArguments.toTypedArray())
-        assertThat(parseResult.matchedOption("output-file").getValue<File>().name).isEqualTo("sampleOutputFile")
+        assertThat(parseResult.matchedOption("output-file").getValue<String>()).isEqualTo("sampleOutputFile")
         assertThat(parseResult.matchedOption("path-separator").getValue<Char>()).isEqualTo('/')
         assertThat(parseResult.matchedPositional(0).getValue<String>()).isEqualTo("sampleFile.cc.json")
     }