+++ analysis/import/MetricGardenerImporter/src/test/kotlin/de/maibornwolff/codecharta/importer/metricgardenerimporter/MetricGardenerImporterTest.kt
@@ -9,13 +9,12 @@ import java.io.File
 class MetricGardenerImporterTest {
 
     private val metricGardenerImporter = MetricGardenerImporter()
-
-    @Test
+     @Test
     fun `should create json uncompressed file`() {
         main(
             arrayOf(
                 "src/test/resources/metricgardener-analysis.json", "-nc",
-                "-o=src/test/resources/metricgardener-analysis.cc.json"
+                "-o=src/test/resources/metricgardener-analysis"
                    )
             )
         val file = File("src/test/resources/metricgardener-analysis.cc.json")
@@ -29,7 +28,7 @@ class MetricGardenerImporterTest {
         main(
             arrayOf(
                 "src/test/resources/metricgardener-analysis.json",
-                "-o=src/test/resources/metricgardener-analysis.cc.json"
+                "-o=src/test/resources/metricgardener-analysis"
                    )
             )
         val file = File("src/test/resources/metricgardener-analysis.cc.json.gz")
@@ -42,9 +41,21 @@ class MetricGardenerImporterTest {
     fun `should create no file when the input file was not specified`() {
         main(
             arrayOf(
-                "-o=src/test/resources/metricgardener-analysis.cc.json"
+                "-o=src/test/resources/metricgardener-analysis.json"
                    )
             )
+        val file = File("src/test/resources/metricgardener-analysis.cc.json.gz")
+        file.deleteOnExit()
+        metricGardenerImporter.call()
+        assertFalse(file.exists())
+    }
+    @Test
+    fun `should create no file, if no output filename is specified`() {
+        main(
+                arrayOf(
+                        "src/test/resources/metricgardener-analysis.json",
+                        )
+        )
         val file = File("src/test/resources/metricgardener-analysis.cc.json.gz")
         file.deleteOnExit()
         metricGardenerImporter.call()